Pass one of moving unaccounted for fragmented or empty clusters to the front of the disk.
Did verify checksums of "copied from" and "copied to" clusters to be equal, one reason
I decided not to even attempt trying to move into clusters that were just evicted. Even with
unit tests there would be too much risk of a catastrophic bug overwriting everything with
garbage thereby ruining the whole experiment in salvaging fragmented files in a FAT without
the FAT. (Itself due to a different but equally catastrophic form of user error, namely
passing the wrong device to dd if=/dev/zero! So I ain't ever gonna be
too careful after that.